Move 'ids' C API module to 'ids_core' module
Rename the entire existing module to 'ids_core' and move it to a
subdirectory.  The actual 'ids' module is pure Python.  For now, it
simply imports all of ids_core.

In the future, ids_core will be the low-level module, and will be
wrapped by ids, which will provide nicer functionality.