Replace ids_core.Camera.start/stop_continuous()
Use the attribute ids_core.Camera.continuous_capture to both
enable/disable capture (with True/False), or query the status, by
getting the value.

This allows the color_mode setter to get the actual status of continuous
capture, rather than attempting to keep track of it with the
start/stop_continuous() methods.