Lookup camera resolution in ids_core.Camera
During ids_core.Camera initialization, look up the sensor info to
determine the width and height of images from the camera.  This is done
in ids_core.Camera, as there is no way to get this information until the
camera has been opened, and the width and height are too important to
simply require ids.Camera to look up and set the values ex post facto.

Since this requires calling ids_core_Camera_getinfo(), initialize the
camera name as well, making it no longer lazily initialized.

Closes #99